Got pixi app embedded in Game component
